+++ src/server/ScriptServer.ts
@@ -4,7 +4,8 @@ import {
 	createConnection, IConnection,
 	TextDocuments, InitializeResult,
 	WorkspaceFolder,
-	TextDocumentSyncKind
+	TextDocumentSyncKind,
+	Disposable
 } from 'vscode-languageserver';
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import { getIndexes, activateIndexes } from './scriptServer/features/indexes';
@@ -23,9 +24,10 @@ const documents = new TextDocuments(TextDocument);
 documents.listen(connection);
 
 let workspaceFolders: WorkspaceFolder[] = [];
+let diagnostics : Disposable | undefined;
+let isDiagnosticsEnabled = true;
 
-connection.onInitialized(() => {
-
+connection.onInitialized((params) => {
 	connection.workspace.onDidChangeWorkspaceFolders((event) => {
 		connection.workspace.getWorkspaceFolders().then(_workspaceFolders => {
 			workspaceFolders = _workspaceFolders || [];
@@ -37,14 +39,39 @@ connection.onInitialized(() => {
 		connection.console.log('Workspace folder change event received');
 	});
 
-	import('./scriptServer/features/gotoLocation').then(gotoLocation => {
-		gotoLocation.activate(connection, documents);
-	});
-	import('./scriptServer/features/competitions').then(competitions => {
-		competitions.activate(connection, documents);
-	});
-	import('./scriptServer/features/diagnostics').then(linting => {
-		linting.activate(connection, documents);
+
+	Promise.all([
+		getIndexes().properties,
+		getIndexes().controllers,
+		getIndexes().scriptFiles,
+		getIndexes().templates]
+	).then(() => {
+		import('./scriptServer/features/gotoLocation').then(gotoLocation => {
+			gotoLocation.activate(connection, documents);
+		});
+		import('./scriptServer/features/competitions').then(competitions => {
+			competitions.activate(connection, documents);
+		});
+
+		if (isDiagnosticsEnabled) {
+			import('./scriptServer/features/diagnostics').then(linting => {
+				diagnostics = linting.activate(connection, documents);
+			});
+		}
+		connection.onDidChangeConfiguration(param => {
+			const disable = param?.settings?.extension?.prophet?.script?.server?.disable?.diagnostics ?? false;
+
+			if ((!disable) !== isDiagnosticsEnabled) {
+				isDiagnosticsEnabled = (!disable);
+				if (disable) {
+					diagnostics?.dispose();
+				} else {
+					import('./scriptServer/features/diagnostics').then(linting => {
+						diagnostics = linting.activate(connection, documents);
+					});
+				}
+			}
+		});
 	});
 });
 
@@ -52,13 +79,9 @@ connection.onInitialized(() => {
 
 connection.onInitialize((params): InitializeResult => {
 
-	connection.console.log('Script Server init...' + JSON.stringify(params.workspaceFolders));
+	connection.console.log('Script Server init... \n' + JSON.stringify(params.workspaceFolders));
 
-	// The VS Code htmlhint settings have changed. Revalidate all documents.
-	// connection.onDidChangeConfiguration((args) => {
-	// 	onDidChangeConfiguration(connection, documents, args);
-	// });
-
+	isDiagnosticsEnabled = !(params.initializationOptions?.disableDiagnostics ?? false);
 
 	workspaceFolders = params.workspaceFolders || [];
 
@@ -118,15 +141,10 @@ connection.onNotification('get.controllers.list', () => {
 });
 
 
-// A text document has changed. Validate the document.
-documents.onDidChangeContent((event) => {
-	// the contents of a text document has changed
-	//validateTextDocument(connection, event.document);
-});
-
 
 // Listen on the connection
 connection.listen();
+
 
 process.once('uncaughtException', err => {
 	console.error(String(err) + '\n' + err.stack);